requirements.txt
pycryptodome>=3.20.0
argparse>=1.4.0